+++ src/utility/index.js
@@ -1,0 +1,7 @@
+import cleanFields from './clean-fields'
+import docToModifier from './doc-to-modifier'
+import cleanNulls from './clean-nulls'
+import reportNulls from './report-nulls'
+import replaceIndexKeys from './replace-index-keys'
+
+export {cleanFields, docToModifier, cleanNulls, reportNulls, replaceIndexKeys}