+++ src/utility/replace-index-keys.js
@@ -1,0 +1,3 @@
+export default function (key) {
+  return key.replace(/\.[0-9]+\./g, '.$.')
+}